move to NodeTypePool.Serializer
